refactor: remove model package
