+++ model.go
@@ -1,4 +1,4 @@
-package model
+package selfish
 
 import (
 	"fmt"
@@ -9,7 +9,20 @@ import (
 
 	"github.com/google/go-github/github"
 	"github.com/pkg/errors"
+	"github.com/podhmo/selfish/pkg/commithistory"
 )
+
+type Commit = commithistory.Commit
+
+// NewCommit creates and initializes a new Commit object.
+func NewCommit(g *github.Gist, alias string, action string) *Commit {
+	return &Commit{
+		ID:        *g.ID,
+		CreatedAt: *g.CreatedAt,
+		Alias:     alias,
+		Action:    action,
+	}
+}
 
 type Gist = github.Gist
